@@ -1,10 +1,12 @@
 export type EventType =
+  | "ON-CLOSE"
   | "ON-NEW-FILE-MENU"
   | "ON-OPEN-FILE-MENU"
   | "ON-SAVE-MENU"
   | "ON-SAVE-AS-MENU";
 
 export type EventOptions<T extends EventType> = { type: T } & {
+  "ON-CLOSE": {};
   "ON-NEW-FILE-MENU": {};
   "ON-OPEN-FILE-MENU": {};
   "ON-SAVE-MENU": {};
@@ -12,9 +14,13 @@ export type EventOptions<T extends EventType> = { type: T } & {
 }[T];
 
 export type EventEmitter = {
-  <T extends EventType>(options: EventOptions<T>): void;
+  <T extends EventType>(options: EventOptions<T>): Promise<void>;
 };
 
 export type EventHandler<T extends EventType> = {
-  (options: EventOptions<T>): void;
+  (options: EventOptions<T>): Promise<void>;
+};
+
+export type UseEventHandler = {
+  <T extends EventType>(eventType: T, handler: EventHandler<T>): void;
 };
