@@ -1,35 +1,30 @@
 import { describe, it } from "mocha";
 import * as assert from "assert";
 import * as moment from "moment";
-import { View, ViewSerializer } from "../../src/persistence/view";
-
-export function assertViewProperties(
-  view: View,
-  id: number,
-  userId: number,
-  bookId: number,
-  date: number
-): void {
-  assert.strictEqual(view.id, id);
-  assert.strictEqual(view.userId, userId);
-  assert.strictEqual(view.bookId, bookId);
-  assert.strictEqual(view.date, date);
-}
+import { View, ViewSerializer } from "../../../src/common/persistence/view";
 
 describe("View", function() {
+  describe("equals", function() {
+    it("Simple Equality", async function() {
+      const a = new View(10823, 35, 707, moment.utc("20200103").valueOf());
+      const b = new View(10823, 35, 707, moment.utc("20200103").valueOf());
+      assert(a.equals(b));
+      assert(b.equals(a));
+    });
+    it("Simple Inequality", async function() {
+      const a = new View(10823, 35, 707, moment.utc("20200103").valueOf());
+      const b = new View(20391, 45, 808, moment.utc("20181230").valueOf());
+      assert(!a.equals(b));
+      assert(!b.equals(a));
+    });
+  });
   describe("ViewSerializer", function() {
     it("Single View Serialization and Deserialization", function() {
       const view = new View(1, 2, 3, 1318781875806);
       const viewSerializer = new ViewSerializer();
       const str = viewSerializer.serialize(view);
       const deserialized = viewSerializer.deserialize(str);
-      assertViewProperties(
-        deserialized,
-        view.id,
-        view.userId,
-        view.bookId,
-        view.date
-      );
+      assert(deserialized.equals(view));
     });
 
     it("Date Serialization", function() {
